Use the new Task-based functions for Stream & Node
Currently based on the `next` branches for the node and core package
where sending to a Stream and exiting node are tasks instead of commands
for composability. This lets me ensure things are executed in order,
fixing bugs like exiting before we show the cursor again.

The node and core packages should be published in June, then this can be
part of a gren-tui 2.0 release.