Fix bugs with a complete rendering refactor
Decouple view from update. Now renders 60 fps like a game loop.

It was too complicated to control the sequencing of update and view,
especially on init when I need to make sure I get cursor position and
clear that sequence before I let the app do anything. This was causing
weird bugs like clearing the wrong thing if the app triggered updates
with an init command.